Fix #2910: Move crd-generator tests out of kubernetes-itests
+ Move crd-generator tests from kubernetes-itests -> kubernetes-tests
