+++ kubernetes-itests/pom.xml
@@ -38,10 +38,6 @@
     </dependency>
     <dependency>
       <groupId>io.fabric8</groupId>
-      <artifactId>crd-generator-apt</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.fabric8</groupId>
       <artifactId>openshift-client</artifactId>
       <scope>test</scope>
     </dependency>
