+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/BaseSpec.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd;
+package io.fabric8.crd.generator;
 
 /**
  * @author <a href="claprun@redhat.com">Christophe Laprun</a>
  */
-public class FooStatus extends BaseStatus {
+public class BaseSpec {
 
 }
