+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/Foo.java
@@ -13,16 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd;
+package io.fabric8.crd.generator;
 
 import io.fabric8.kubernetes.api.model.Namespaced;
-import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Plural;
+import io.fabric8.kubernetes.model.annotation.Version;
 
-/**
- * @author <a href="claprun@redhat.com">Christophe Laprun</a>
- */
-public abstract class Base<Spec extends BaseSpec, Status extends BaseStatus>
-  extends CustomResource<Spec, Status>
-  implements Namespaced {
+@Version("v1alpha1")
+@Group("acme.com")
+public class Foo extends Base<FooSpec, FooStatus> implements Namespaced {
+
 
 }
